Add run_namd function
This function is useful to perform the simulations. If simulations fails it restart the calculation until 5 times more. If it still fails, then it writes an error message but continues the other calculations. It is useful because using certain values for RMSD and RDGYR for some molecules can cause instabilities for intrinsic molecule properties, as size, rigidity, among others.