Update the overall protocol to include smiles reading
This change was performed to allow the smiles support.
This change makes more difficult to read the overall protocol, however it provides the necessary support for smiles. In this case a smi file with one or more entries is read and a new folder is created to process every entry.