Add the clustering stage to the protocol
This change implements a clustering of the structures generated in the sampling stage. A distance matrix is created using the rmsd between all the dcd files. Then, the centroids are written to independent PDB files named: centroid_x.pdb, where x is the id. The number of clusters to generate is controlled by the option: n_confs in the parser file. Defaults=50.