Add a clustering modification
This method performs a modification to the clustering.
The grouping method, create small groups of structures below 3 Angstroms of RMSD.
The subclustering only performs a cluster on one array of structures and the specified n_clusters number.
Then all groups are subclustered according to the variation of the groups with different cuttofs. The remaining structures are clustered to the desired final number of clusters.