Update protocol.cr
Those changes includes the new parameters that could be modified with the input options:
n_variants, threshold_rmsd_variants, spacing_rdgyr_variants and fullsamples. The first is the number of random conformations to use as a reference in every window. The threshold_rmsd_variants is the maximum RMSD that variants should have between them to avoid to repeat similar conformations, this RMSD is adaptive iteratively in the process. The spacing_rdgyr_variants is a ghost option and should be erased.