Update protocol for amber24 version
The bcc charge model was not the right option in conjuction with gaff2 ff according to amber manual. The recommended option is to use gaff2/abcg2. Now we will test that version for large datasets, I hope it works! if not... weird, we will go back to use the non-recommended option.